requirements.txt
cx_Oracle==8.3.0
oracledb==3.4.0
python-dotenv==1.2.1
